document planned "skip" user action
With this interface, it should be able to capture a large variety of
interactions. All it takes is a nice web interface with JavaScript. I
can imagine that sub-problems appear with more indentation, such as in
little boxes, which then disappear or shrink when the subproblem is
solved. (on RITS action "exit").