ENHANCED: Implement the "again" LORITS element for repeating the task.
"again" is an abbreviation for the most recently posed task at the
time the actions are being expanded. Note that sub-tasks that are
spawned when processing these actions therefore have no impact on
this, and that is also the intention. We do not mean to retry the
sub-task, but the task that spawned it.