fold away some LORITS elements
The precise relation between drilling exercises, exams and tutorials
is still to be defined. We may need to selectively display
informational messages. One way is to extend LORITS with an element like:

   help(Pred)

which, when applicable, calls Pred with the (currently) 3 additional
arguments (Task, Answer, History) so that it can print something.